requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
aiofiles==23.2.1
    # via nicegui
aiohttp==3.8.6
    # via nicegui
aiosignal==1.3.1
    # via aiohttp
anyascii==0.3.2
    # via wagtail
anyio==3.7.1
    # via
    #   fastapi
    #   httpcore
    #   starlette
    #   watchfiles
asgiref==3.7.2
    # via django
asttokens==2.4.0
    # via stack-data
async-timeout==4.0.3
    # via aiohttp
attrs==23.1.0
    # via aiohttp
backcall==0.2.0
    # via ipython
beautifulsoup4==4.11.2
    # via wagtail
bidict==0.22.1
    # via python-socketio
black==23.9.1
    # via -r requirements.in
boto3==1.28.62
    # via django-ses
botocore==1.31.62
    # via
    #   boto3
    #   s3transfer
build==1.0.3
    # via pip-tools
certifi==2023.7.22
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.0
    # via
    #   aiohttp
    #   requests
click==8.1.7
    # via
    #   black
    #   pip-tools
    #   uvicorn
coverage[toml]==7.3.2
    # via
    #   -r requirements.in
    #   pytest-cov
crispy-bootstrap4==2022.1
    # via -r requirements.in
cryptography==41.0.4
    # via django-ses
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via willow
django==4.2.6
    # via
    #   -r requirements.in
    #   crispy-bootstrap4
    #   django-crispy-forms
    #   django-extensions
    #   django-filter
    #   django-modelcluster
    #   django-permissionedforms
    #   django-salesman
    #   django-ses
    #   django-taggit
    #   django-treebeard
    #   djangorestframework
    #   model-bakery
    #   wagtail
django-crispy-forms==2.0
    # via crispy-bootstrap4
django-environ==0.11.2
    # via -r requirements.in
django-extensions==3.2.3
    # via -r requirements.in
django-filter==23.3
    # via wagtail
django-modelcluster==6.1
    # via wagtail
django-permissionedforms==0.1
    # via wagtail
django-salesman[pygments]==1.2.0
    # via -r requirements.in
django-ses[events]==3.5.0
    # via -r requirements.in
django-taggit==4.0.0
    # via wagtail
django-treebeard==4.7
    # via wagtail
djangorestframework==3.14.0
    # via
    #   django-salesman
    #   wagtail
draftjs-exporter==2.1.7
    # via wagtail
et-xmlfile==1.1.0
    # via openpyxl
exceptiongroup==1.1.3
    # via
    #   anyio
    #   ipython
    #   pytest
execnet==2.0.2
    # via pytest-xdist
executing==2.0.0
    # via stack-data
factory-boy==3.3.0
    # via wagtail-factories
faker==19.10.0
    # via factory-boy
fastapi==0.103.2
    # via
    #   fastapi-socketio
    #   nicegui
fastapi-socketio==0.0.10
    # via nicegui
filetype==1.2.0
    # via willow
freezegun==1.2.2
    # via pytest-freezegun
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
gunicorn==21.2.0
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
    #   wsproto
html5lib==1.1
    # via wagtail
htmx==0.0.0
    # via -r requirements.in
httpcore==0.17.3
    # via httpx
httptools==0.6.0
    # via uvicorn
httpx==0.24.1
    # via nicegui
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==6.8.0
    # via build
iniconfig==2.0.0
    # via pytest
ipython==8.16.1
    # via -r requirements.in
itsdangerous==2.1.2
    # via nicegui
jedi==0.19.1
    # via ipython
jinja2==3.1.2
    # via nicegui
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
l18n==2021.3
    # via wagtail
markdown2==2.4.10
    # via nicegui
markupsafe==2.1.3
    # via jinja2
matplotlib-inline==0.1.6
    # via ipython
model-bakery==1.15.0
    # via -r requirements.in
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via black
nicegui==1.3.16
    # via htmx
openpyxl==3.1.2
    # via wagtail
orjson==3.9.8
    # via nicegui
packaging==23.2
    # via
    #   black
    #   build
    #   gunicorn
    #   pytest
parso==0.8.3
    # via jedi
pathspec==0.11.2
    # via black
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==10.0.1
    # via
    #   pillow-heif
    #   wagtail
pillow-heif==0.13.0
    # via willow
pip-tools==7.3.0
    # via -r requirements.in
platformdirs==3.11.0
    # via black
pluggy==1.3.0
    # via pytest
prompt-toolkit==3.0.39
    # via ipython
pscript==0.7.7
    # via vbuild
psycopg2==2.9.9
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pycparser==2.21
    # via cffi
pydantic==1.10.13
    # via
    #   fastapi
    #   htmx
pygments==2.16.1
    # via
    #   django-salesman
    #   ipython
    #   nicegui
pymemcache==4.0.0
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via build
pytest==7.4.2
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-django
    #   pytest-env
    #   pytest-freezegun
    #   pytest-xdist
pytest-cov==4.1.0
    # via -r requirements.in
pytest-django==4.5.2
    # via -r requirements.in
pytest-env==1.0.1
    # via -r requirements.in
pytest-freezegun==0.4.2
    # via -r requirements.in
pytest-xdist==3.3.1
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   botocore
    #   faker
    #   freezegun
python-dotenv==1.0.0
    # via uvicorn
python-engineio==4.7.1
    # via python-socketio
python-multipart==0.0.6
    # via nicegui
python-socketio==5.9.0
    # via fastapi-socketio
pytz==2023.3.post1
    # via
    #   django-modelcluster
    #   django-ses
    #   djangorestframework
    #   l18n
pyyaml==6.0.1
    # via uvicorn
requests==2.31.0
    # via
    #   django-ses
    #   wagtail
ruff==0.0.292
    # via -r requirements.in
s3transfer==0.7.0
    # via boto3
simple-websocket==1.0.0
    # via python-engineio
six==1.16.0
    # via
    #   asttokens
    #   html5lib
    #   l18n
    #   python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
soupsieve==2.5
    # via beautifulsoup4
sqlparse==0.4.4
    # via django
stack-data==0.6.3
    # via ipython
starlette==0.27.0
    # via fastapi
telepath==0.3.1
    # via wagtail
tomli==2.0.1
    # via
    #   black
    #   build
    #   coverage
    #   pip-tools
    #   pyproject-hooks
    #   pytest
traitlets==5.11.2
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.8.0
    # via
    #   asgiref
    #   black
    #   fastapi
    #   ipython
    #   nicegui
    #   pydantic
    #   starlette
urllib3==1.26.17
    # via
    #   botocore
    #   requests
uvicorn[standard]==0.22.0
    # via nicegui
uvloop==0.17.0
    # via uvicorn
vbuild==0.8.2
    # via nicegui
wagtail==5.1.2
    # via
    #   -r requirements.in
    #   wagtail-factories
wagtail-factories==4.1.0
    # via -r requirements.in
watchfiles==0.20.0
    # via
    #   nicegui
    #   uvicorn
wcwidth==0.2.8
    # via prompt-toolkit
webencodings==0.5.1
    # via html5lib
websockets==11.0.3
    # via uvicorn
wheel==0.41.2
    # via pip-tools
whitenoise==6.6.0
    # via -r requirements.in
willow[heif]==1.6.2
    # via wagtail
wsproto==1.2.0
    # via simple-websocket
yarl==1.9.2
    # via aiohttp
zipp==3.17.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
